move .js and .css files out of the root
Move twinkle.js, morebits.js, CSS files, etc. out of the root. The root directory should be config files and readmes only.

Wikimedia also does this. JS files are always all placed in a directory called /modules/ (or sometimes a subdirectory with a different name, the other name escpaes me at the moment). But never in the root.

Fixes #2078

To test:
- npm start
- perl deploy.pl

Maybe:
- put the old modules in a subdirectory? modules/menuItems/*.js or something?
