+++ jest.config.js
@@ -5,8 +5,6 @@ module.exports = {
 	testEnvironment: 'jsdom',
 	setupFilesAfterEnv: ['mock-mediawiki', '<rootDir>/tests/jest.setup.js'],
 	collectCoverageFrom: [
-		'morebits.js',
-		'twinkle.js',
 		'modules/**/*.{js,jsx,ts,tsx}'
 	]
 };
