+++ modules/morebits.js
