+++ scripts/dev-server.js
@@ -14,8 +14,8 @@ async function readFiles(filePaths) {
 }
 const server = http.createServer(async (request, response) => {
 	const moduleFiles = (await fs.readdir('./modules')).filter(f => f.endsWith('.js'));
-	const jsFiles = ['morebits.js', 'twinkle.js'].concat(moduleFiles.map(f => 'modules/' + f));
-	const cssFiles = ['morebits.css', 'twinkle.css'];
+	const jsFiles = moduleFiles.map(f => 'modules/' + f);
+	const cssFiles = ['modules/morebits.css', 'modules/twinkle.css'];
 
 	let jsCode = `mw.loader.using(['jquery.ui', 'ext.gadget.select2']).then(function () {\n`;
